requirements.txt
amqp==5.2.0
anyio==4.0.0
anytree==2.9.0
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
beautifulsoup4==4.12.2
billiard==4.2.0
celery==5.3.6
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==41.0.2
defusedxml==0.7.1
dhanhq==1.2.2
Django==4.2.2
django-4-jet==1.0.9
django-adminlte-3==0.1.6
django-allauth==0.54.0
django-ash==0.0.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-crontab==0.7.1
django-jet-reboot==1.3.3
django-timezone-field==6.1.0
exceptiongroup==1.1.3
frozendict==2.3.8
fyers-apiv2==2.0.6
google-api-python-client==1.4.1
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httplib2==0.22.0
httptools==0.6.1
idna==2.10
kombu==5.3.4
loguru==0.7.0
lxml==4.9.3
numpy==1.25.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.2
pandas==2.0.2
paper-admin==7.1.1
pbkdf2==1.3
Pillow==10.0.0
prompt-toolkit==3.0.43
py5paisa==0.7.1
pya3==1.0.29
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.7.0
pyotp==2.8.0
pyparsing==3.1.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
razorpay==1.3.0
redis==5.0.1
rel==0.4.9
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
smartapi-python==1.3.0
sniffio==1.3.0
soupsieve==2.5
tblib==3.0.0
tornado==6.1
typing_extensions==4.6.3
tzdata==2023.3
upstox-python-sdk==2.3.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.24.0.post1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==1.2.1
websockets==12.0
